move sshd container configuration to docker-entrypoint.sh, production db drop safegard, event creation page fix, move github pip dependency to setyp,py, put relevant mkdirs in prestart.sh, add VOLUME in DOckerfile

+++ web/app/ponthe/cli.py
@@ -16,6 +16,9 @@ from .services import UserService, GalleryService
 
 def _drop_and_recreate_db():
     click.echo("Emptying database...")
+    if os.environ.get("TAG") == "master":
+        app.logger.error("Environment variable TAG=master detected: you are on a production environment, database drop aborted")
+        exit()
     db.drop_all()
     db.create_all()
